requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
pyyaml==6.0.1
streamlit==1.29.0
plotly==5.18.0
